feat: restructure components in feedback folder
