+++ packages/components/src/components/feedback/Spinner/index.ts
