+++ packages/components/src/components/feedback/index.ts
@@ -1,0 +1,6 @@
+export * from './Alert';
+export * from './InlineAlert';
+export * from './ProgressBar';
+export * from './ProgressCircular';
+export * from './Spinner';
+export * from './Toast';
