+++ packages/components/src/components/index.ts
@@ -1,3 +1,3 @@
 export * from './layouts';
 export * from './primitives';
-export * from './notifications';
+export * from './feedback';
