--- packages/components/src/components/notifications/index.ts
@@ -1,3 +1,0 @@
-export * from './Alert';
-export * from './InlineAlert';
-export * from './Toast';
