+++ packages/components/src/components/primitives/index.ts
@@ -17,13 +17,10 @@ export * from './Modal';
 export * from './NumberInput';
 export * from './NumberedList';
 export * from './PinInput';
-export * from './ProgressBar';
-export * from './ProgressCircular';
 export * from './Radio';
 export * from './RangeSlider';
 export * from './Select';
 export * from './Slider';
-export * from './Spinner';
 export * from './Switch';
 export * from './Symbol';
 export * from './Tabs';
